+++ ginga/examples/gw/example1_video.py
@@ -37,6 +37,7 @@ responsive to user actions.
 import sys
 import time
 import threading
+from argparse import ArgumentParser
 
 import numpy as np
 
@@ -271,14 +272,8 @@ def main(options, args):
 
 if __name__ == '__main__':
 
-    # Parse command line options
-    from argparse import ArgumentParser
-
     argprs = ArgumentParser()
 
-    argprs.add_argument("--debug", dest="debug", default=False,
-                        action="store_true",
-                        help="Enter the pdb debugger on main()")
     argprs.add_argument("--fps", dest="fps", metavar="FPS",
                         type=float, default=None,
                         help="Force a FPS (frames/sec)")
@@ -288,29 +283,10 @@ if __name__ == '__main__':
     argprs.add_argument("-t", "--toolkit", dest="toolkit", metavar="NAME",
                         default='qt',
                         help="Choose GUI toolkit (gtk|qt)")
-    argprs.add_argument("--profile", dest="profile", action="store_true",
-                        default=False,
-                        help="Run the profiler on main()")
     argprs.add_argument("-r", "--render", dest="render", default='widget',
                         help="Set render type {widget|opengl}")
     log.addlogopts(argprs)
 
     (options, args) = argprs.parse_known_args(sys.argv[1:])
 
-    # Are we debugging this?
-    if options.debug:
-        import pdb
-
-        pdb.run('main(options, args)')
-
-    # Are we profiling this?
-    elif options.profile:
-        import profile
-
-        print(("%s profile:" % sys.argv[0]))
-        profile.run('main(options, args)')
-
-    else:
-        main(options, args)
-
-# END
+    main(options, args)